+++ R/extract.r
@@ -24,10 +24,15 @@
 #' # Simplify results into character matrix
 #' str_extract_all(shopping_list, "\\b[a-z]+\\b", simplify = TRUE)
 #' str_extract_all(shopping_list, "\\d", simplify = TRUE)
+#'
+#' # Extract all words
+#' str_extract_all("This is, suprisingly, a sentence.", boundary("word"))
 str_extract <- function(string, pattern) {
   switch(type(pattern),
-    empty = ,
-    bound = stop("Not implemented", call. = FALSE),
+    empty = stri_extract_first_boundaries(string, pattern,
+      opts_brkiter = stri_opts_brkiter(type = "character")),
+    bound = stri_extract_first_boundaries(string, pattern,
+      opts_brkiter = attr(pattern, "options")),
     fixed = stri_extract_first_fixed(string, pattern,
       opts_fixed = attr(pattern, "options")),
     coll  = stri_extract_first_coll(string, pattern,
@@ -41,13 +46,19 @@ str_extract <- function(string, pattern) {
 #' @export
 str_extract_all <- function(string, pattern, simplify = FALSE) {
   switch(type(pattern),
-    empty = ,
-    bound = stop("Not implemented", call. = FALSE),
+    empty = stri_extract_all_boundaries(string, pattern,
+      simplify = simplify, omit_no_match = TRUE,
+      opts_brkiter = stri_opts_brkiter(type = "character")),
+    bound = stri_extract_all_boundaries(string, pattern,
+      simplify = simplify, omit_no_match = TRUE,
+      opts_brkiter = attr(pattern, "options")),
     fixed = stri_extract_all_fixed(string, pattern,
       opts_fixed = attr(pattern, "options")),
     coll  = stri_extract_all_coll(string, pattern,
-      simplify = simplify, omit_no_match = TRUE, attr(pattern, "options")),
+      simplify = simplify, omit_no_match = TRUE,
+      opts_collator = attr(pattern, "options")),
     regex = stri_extract_all_regex(string, pattern,
-      simplify = simplify, omit_no_match = TRUE, attr(pattern, "options"))
+      simplify = simplify, omit_no_match = TRUE,
+      opts_regex = attr(pattern, "options"))
   )
 }